front_refactor : cart, comparison, 함수

+++ front-end/queant/src/store/modules/portfolio.js
@@ -4,18 +4,22 @@ import axios from 'axios'
 
 export default {
   state: {
+    product_name: '',
     portfolio: [],
     customProducts: [],
     portfolios: [],
     comparisonPortfolio: [],
     newlyAddedPortfolio: [],
-    deletedPortfolio:[]
+    deletedPortfolio:[],
     },
   getters: {
+    product_name: state => state.product_name,
     portfolio: state => state.portfolio,
     customProducts: state => state.customProducts,
     portfolios: state => state.portfolios,
-    comparisonPortfolio: state => state.comparisonPortfolio
+    comparisonPortfolio: state => state.comparisonPortfolio,
+    newlyAddedPortfolio: state => state.newlyAddedPortfolio,
+    deletedPortfolio: state => state.deletedPortfolio,
   },
   mutations: {
     SET_PORTFOLIO: (state, portfolio) => state.portfolio = portfolio,
@@ -34,12 +38,16 @@ export default {
         console.log('내 포트폴리오에 상품을 추가했습니다.')
       }
     },
+    CLEAR_CPORTFOLIO_DB(state) {
+      state.newlyAddedPortfolio = [],
+      state.deletedPortfolio = []
+    },
 
     POP_CPORTFOLIO(state, cportfolio_cnt) {
       state.deletedPortfolio.push(cportfolio_cnt)
       let portfolioNo = cportfolio_cnt
       // 삭제된 포트폴리오의 이후 포트폴리오의 cnt를 1씩 차감
-      for (var i=portfolioNo; i <= state.comparisonPortfolio.length-1; i++) {
+      for (let i=portfolioNo; i <= state.comparisonPortfolio.length-1; i++) {
         state.comparisonPortfolio[i].cportfolio_cnt -= 1
         console.log(state.comparisonPortfolio[i])
       }
@@ -69,10 +77,20 @@ export default {
       let portfolioNo = value[0]
       let product = value[1]
       let cportfolios = state.comparisonPortfolio
-      if (cportfolios[portfolioNo-1].products.find(cportfolioItem => cportfolioItem.product.product_id === product.product.product_id)) {
+      let pushproduct = {
+        portfolio_no: portfolioNo,
+        product_id: product.product.product_id,
+        amount: product.amount,
+        condition_ids: product.condition_ids,
+        option_id: product.option_id,
+        start_date: product.start_date,
+        end_date: product.end_date,
+        name: product.product.name
+      }
+      if (cportfolios[portfolioNo-1].products.find(cportfolioItem => cportfolioItem.product_id === product.product.product_id)) {
         alert('이미 포트폴리오에 있는 상품입니다.')
       } else {
-        state.comparisonPortfolio[portfolioNo-1].products.push(product)
+        state.comparisonPortfolio[portfolioNo-1].products.push(pushproduct)
       console.log(`${portfolioNo}번 포트폴리오에 상품이 담겼습니다.`)
       }
     },
@@ -105,6 +123,9 @@ export default {
 
       state.portfolios[portfolioIdx].splice(productIdx, 1)
       console.log(`${portfolioIdx+1}번 포트폴리오의 ${productIdx+1}번 상품을 삭제했습니다.`)
+    },
+    CHANGE_NAMES(state, name) {
+      state.product_name = name
     }
   },
   actions: {
@@ -261,35 +282,38 @@ export default {
         console.log(err)
       })
     },
-    saveToDb({ getters, state }) {
-      //삭제된 포트폴리오 DB에서 제거하기
-      for (const port_no of state.deletedPortfolio) {
-        axios({
-          url: spring.portfolio.portfolio(),
-          method: 'delete',
-          data: {
-            member_id: getters.userInfo.member_id,
-            portfolio_no:port_no
-          }
-        }).then(res => {
-          console.log(res)
-          console.log(port_no+'번 포트폴리오가 DB에서 제거되었습니다.')
-        })
-        .catch(err => {
-          console.log(err)
-        })
-      }
 
+    // 가상 포트폴리오 DB 저장 및 가져오기
+    saveToDb({ dispatch, state, getters }) {
+      //삭제된 포트폴리오 DB에서 제거하기
+      if (state.deletedPortfolio.length != 0) {
+        for (const port_no of state.deletedPortfolio) {
+          axios({
+            url: spring.portfolio.portfolio(),
+            method: 'delete',
+            data: {
+              member_id: getters.userInfo.member_id,
+              portfolio_no:port_no
+            }
+          }).then(res => {
+            console.log(res)
+            console.log(port_no+'번 포트폴리오가 DB에서 제거되었습니다.')
+          })
+          .catch(err => {
+            console.log(err)
+          })
+        }
+      }
       //현재 남아있는 포트폴리오 수정 및 삽입
-      for (var i = 1; i <= state.comparisonPortfolio.length; i++){
+      for (let i = 1; i <= state.comparisonPortfolio.length; i++){
         if (state.newlyAddedPortfolio.includes(i)) {
           console.log(state.comparisonPortfolio[i-1])
-          var tempPortfolioList = [];
+          let tempPortfolioList = [];
           for (const port of state.comparisonPortfolio[i - 1].products) {
             console.log(port.option_id)
-            var temp = {
+            let temp = {
               portfolio_no:i,
-              product_id:port.product.product_id,
+              product_id:port.product_id,
               amount:port.amount,
               condition_ids: port.condition_ids,
               option_id:port.option_id,
@@ -299,6 +323,7 @@ export default {
             tempPortfolioList.push(temp)
           }
           console.log(tempPortfolioList)
+          console.log(getters.userInfo)
           axios({
             url: spring.portfolio.portfolio(),
             method: 'post',
@@ -314,13 +339,13 @@ export default {
             console.log(err)
           })
         } else {
-          var tempPortfolioList = [];
+          let tempPortfolioList = [];
           console.log(state.comparisonPortfolio[i-1])
           for (const port of state.comparisonPortfolio[i - 1].products) {
             console.log(port.option_id)
-            var temp = {
+            let temp = {
               portfolio_no:i,
-              product_id:port.product.product_id,
+              product_id:port.product_id,
               amount:port.amount,
               condition_ids: port.condition_ids,
               option_id: port.option_id,
@@ -349,9 +374,12 @@ export default {
       }
       
     },
-    getFromDb({ getters, commit}) {
+    getFromDb({ commit, dispatch, state, getters, actions }) {
+      // Clear DB를 실행(추가, 삭제 초기화)
+      commit('CLEAR_CPORTFOLIO_DB');
+      // 포트폴리오 초기화
       commit('CLEAR_CPORTFOLIOS');
-
+      // 정보 받아오기
       axios({
         url: spring.portfolio.virtual(),
         method: 'post',
@@ -361,12 +389,33 @@ export default {
       })
       .then(res => {
         console.log(res.data)
-        const dbPortfolio = res.data;
-
-        commit('UPDATE_CPORTFOLIO_FROM_DB', dbPortfolio);
-
+        const firstdata = res.data
+        let bringdata = []
+        // 넣을 데이터를 생성
+        for (let i = 0; i<firstdata.length; i++) {
+          bringdata.push({
+            cportfolio_cnt: i+1,
+            products: []
+          })
+        }
         
-
+        // 받아온 데이터를 다시 형식에 맞게 변경
+        for (let i = 0; i<firstdata.length; i++) {
+          for (let j = 0; j<firstdata[i].length; j++) {
+            bringdata[i].products.push({
+              portfolio_no: firstdata[i][j].portfolio_no,
+              product_id: firstdata[i][j].product_id,
+              amount: firstdata[i][j].amount,
+              condition_ids: firstdata[i][j].condition_ids,
+              option_id: firstdata[i][j].option_id,
+              start_date: firstdata[i][j].start_date,
+              end_date: firstdata[i][j].end_date,
+              name: firstdata[i][j].product.name
+            })
+          }
+        }
+        // ComparisonPortfolio를 업데이트
+        commit('UPDATE_CPORTFOLIO_FROM_DB', bringdata);
       })
       .catch(err => {
         console.log(err)
